# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Family Tree API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/trees": {
      "get": {
        "tags": [
          "trees"
        ],
        "summary": "List Trees",
        "operationId": "list_trees_api_trees_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TreeListItem"
                  },
                  "type": "array",
                  "title": "Response List Trees Api Trees Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "trees"
        ],
        "summary": "Create Tree",
        "operationId": "create_tree_api_trees_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TreeCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TreeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/trees/{tree_id}": {
      "get": {
        "tags": [
          "trees"
        ],
        "summary": "Get Tree",
        "operationId": "get_tree_api_trees__tree_id__get",
        "parameters": [
          {
            "name": "tree_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tree Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TreeDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "trees"
        ],
        "summary": "Update Tree",
        "operationId": "update_tree_api_trees__tree_id__patch",
        "parameters": [
          {
            "name": "tree_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tree Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TreeUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TreeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "trees"
        ],
        "summary": "Delete Tree",
        "operationId": "delete_tree_api_trees__tree_id__delete",
        "parameters": [
          {
            "name": "tree_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tree Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/trees/{tree_id}/subtrees": {
      "get": {
        "tags": [
          "subtrees"
        ],
        "summary": "List Subtrees",
        "operationId": "list_subtrees_api_trees__tree_id__subtrees_get",
        "parameters": [
          {
            "name": "tree_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tree Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubtreeOut"
                  },
                  "title": "Response List Subtrees Api Trees  Tree Id  Subtrees Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "subtrees"
        ],
        "summary": "Create Subtree",
        "operationId": "create_subtree_api_trees__tree_id__subtrees_post",
        "parameters": [
          {
            "name": "tree_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tree Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubtreeCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubtreeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/trees/{tree_id}/subtrees/{subtree_id}": {
      "patch": {
        "tags": [
          "subtrees"
        ],
        "summary": "Update Subtree",
        "operationId": "update_subtree_api_trees__tree_id__subtrees__subtree_id__patch",
        "parameters": [
          {
            "name": "tree_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tree Id"
            }
          },
          {
            "name": "subtree_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Subtree Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubtreeUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubtreeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "subtrees"
        ],
        "summary": "Delete Subtree",
        "operationId": "delete_subtree_api_trees__tree_id__subtrees__subtree_id__delete",
        "parameters": [
          {
            "name": "tree_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tree Id"
            }
          },
          {
            "name": "subtree_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Subtree Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/trees/{tree_id}/subtrees/{subtree_id}/people": {
      "get": {
        "tags": [
          "subtrees"
        ],
        "summary": "Subtree People",
        "operationId": "subtree_people_api_trees__tree_id__subtrees__subtree_id__people_get",
        "parameters": [
          {
            "name": "tree_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tree Id"
            }
          },
          {
            "name": "subtree_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Subtree Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PersonOut"
                  },
                  "title": "Response Subtree People Api Trees  Tree Id  Subtrees  Subtree Id  People Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/trees/{tree_id}/people": {
      "get": {
        "tags": [
          "people"
        ],
        "summary": "List People",
        "operationId": "list_people_api_trees__tree_id__people_get",
        "parameters": [
          {
            "name": "tree_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tree Id"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PersonOut"
                  },
                  "title": "Response List People Api Trees  Tree Id  People Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "people"
        ],
        "summary": "Create Person",
        "operationId": "create_person_api_trees__tree_id__people_post",
        "parameters": [
          {
            "name": "tree_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tree Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PersonCreate": {
        "properties": {
          "given_names": {
            "type": "string",
            "maxLength": 200,
            "title": "Given Names",
            "default": ""
          },
          "surname": {
            "type": "string",
            "maxLength": 200,
            "title": "Surname",
            "default": ""
          },
          "birth_surname": {
            "type": "string",
            "maxLength": 200,
            "title": "Birth Surname",
            "default": ""
          },
          "nickname": {
            "type": "string",
            "maxLength": 200,
            "title": "Nickname",
            "default": ""
          },
          "native_name": {
            "type": "string",
            "maxLength": 400,
            "title": "Native Name",
            "default": ""
          },
          "sex": {
            "$ref": "#/components/schemas/Sex",
            "default": "unknown"
          },
          "bio": {
            "type": "string",
            "title": "Bio",
            "default": ""
          },
          "occupation": {
            "type": "string",
            "maxLength": 200,
            "title": "Occupation",
            "default": ""
          },
          "nationality": {
            "type": "string",
            "maxLength": 200,
            "title": "Nationality",
            "default": ""
          },
          "education": {
            "type": "string",
            "maxLength": 300,
            "title": "Education",
            "default": ""
          },
          "links": {
            "items": {
              "$ref": "#/components/schemas/SocialLink-Input"
            },
            "type": "array",
            "maxItems": 12,
            "title": "Links"
          },
          "vehicles": {
            "items": {
              "type": "string",
              "maxLength": 120,
              "minLength": 1
            },
            "type": "array",
            "maxItems": 12,
            "title": "Vehicles"
          },
          "pets": {
            "items": {
              "$ref": "#/components/schemas/Pet-Input"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Pets"
          },
          "favorites": {
            "items": {
              "$ref": "#/components/schemas/Favorite"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Favorites"
          },
          "is_living": {
            "type": "boolean",
            "title": "Is Living",
            "default": true
          },
          "is_me": {
            "type": "boolean",
            "title": "Is Me",
            "default": false
          },
          "relative": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RelativeLink"
              },
              {
                "type": "null"
              }
            ]
          },
          "birth": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FuzzyDate"
              },
              {
                "type": "null"
              }
            ]
          },
          "death": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FuzzyDate"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "PersonCreate"
      },
...
      "PersonOut": {
        "properties": {
          "given_names": {
            "type": "string",
            "maxLength": 200,
            "title": "Given Names",
            "default": ""
          },
          "surname": {
            "type": "string",
            "maxLength": 200,
            "title": "Surname",
            "default": ""
          },
          "birth_surname": {
            "type": "string",
            "maxLength": 200,
            "title": "Birth Surname",
            "default": ""
          },
          "nickname": {
            "type": "string",
            "maxLength": 200,
            "title": "Nickname",
            "default": ""
          },
          "native_name": {
            "type": "string",
            "maxLength": 400,
            "title": "Native Name",
            "default": ""
          },
          "sex": {
            "$ref": "#/components/schemas/Sex",
            "default": "unknown"
          },
          "bio": {
            "type": "string",
            "title": "Bio",
            "default": ""
          },
          "occupation": {
            "type": "string",
            "maxLength": 200,
            "title": "Occupation",
            "default": ""
          },
          "nationality": {
            "type": "string",
            "maxLength": 200,
            "title": "Nationality",
            "default": ""
          },
          "education": {
            "type": "string",
            "maxLength": 300,
            "title": "Education",
            "default": ""
          },
          "links": {
            "items": {
              "$ref": "#/components/schemas/SocialLink-Output"
            },
            "type": "array",
            "maxItems": 12,
            "title": "Links"
          },
          "vehicles": {
            "items": {
              "type": "string",
              "maxLength": 120,
              "minLength": 1
            },
            "type": "array",
            "maxItems": 12,
            "title": "Vehicles"
          },
          "pets": {
            "items": {
              "$ref": "#/components/schemas/Pet-Output"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Pets"
          },
          "favorites": {
            "items": {
              "$ref": "#/components/schemas/Favorite"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Favorites"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tree_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tree Id"
          },
          "is_living": {
            "type": "boolean",
            "title": "Is Living"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "linked_user_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Linked User Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "birth": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VitalOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "death": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VitalOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "given_names",
          "surname",
          "birth_surname",
          "nickname",
          "native_name",
          "sex",
          "bio",
          "occupation",
          "nationality",
          "education",
          "links",
          "vehicles",
          "pets",
          "favorites",
          "id",
          "tree_id",
          "is_living",
          "display_name",
          "linked_user_id",
          "created_at",
          "updated_at",
          "birth",
          "death"
        ],
        "title": "PersonOut"
      },
...
      "Role": {
        "type": "string",
        "enum": [
          "personal",
          "contributor",
          "admin",
          "owner"
        ],
        "title": "Role",
        "description": "Membership roles, lowest to highest privilege."
      },
      "Sex": {
        "type": "string",
        "enum": [
          "male",
          "female",
          "other",
          "unknown"
        ],
        "title": "Sex"
      },
...
      "SocialLink-Output": {
        "properties": {
          "url": {
            "type": "string",
            "maxLength": 500,
            "title": "Url"
          },
          "label": {
            "type": "string",
            "maxLength": 60,
            "title": "Label",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "url",
          "label"
        ],
        "title": "SocialLink"
      },
      "SubtreeCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "root_person_id": {
            "type": "string",
            "format": "uuid",
            "title": "Root Person Id"
          },
          "direction": {
            "$ref": "#/components/schemas/SubtreeDirection",
            "default": "descendants"
          },
          "include_spouses": {
            "type": "boolean",
            "title": "Include Spouses",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "name",
          "root_person_id"
        ],
        "title": "SubtreeCreate"
      },
      "SubtreeDirection": {
        "type": "string",
        "enum": [
          "descendants",
          "ancestors",
          "both"
        ],
        "title": "SubtreeDirection"
      },
      "SubtreeOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "root_person_id": {
            "type": "string",
            "format": "uuid",
            "title": "Root Person Id"
          },
          "direction": {
            "$ref": "#/components/schemas/SubtreeDirection"
          },
          "include_spouses": {
            "type": "boolean",
            "title": "Include Spouses"
          },
          "member_count": {
            "type": "integer",
            "title": "Member Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "root_person_id",
          "direction",
          "include_spouses",
          "member_count"
        ],
        "title": "SubtreeOut"
      },
...
      "SubtreeUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "root_person_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Root Person Id"
          },
          "direction": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SubtreeDirection"
              },
              {
                "type": "null"
              }
            ]
          },
          "include_spouses": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Include Spouses"
          }
        },
        "type": "object",
        "title": "SubtreeUpdate"
      },
...
      "TreeCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "TreeCreate"
      },
      "TreeDetailOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "access": {
            "$ref": "#/components/schemas/MyAccessOut"
          },
          "person_count": {
            "type": "integer",
            "title": "Person Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "created_at",
          "access",
          "person_count"
        ],
        "title": "TreeDetailOut"
      },
      "TreeListItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "highest_role": {
            "$ref": "#/components/schemas/Role"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "created_at",
          "highest_role"
        ],
        "title": "TreeListItem"
      },
      "TreeOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "created_at"
        ],
        "title": "TreeOut"
      },
      "TreeUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "title": "TreeUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}